pyproject.toml
[tool.poetry]
name = "iou_app"
version = "0.13.0"
description = "IOU App"
authors = ["Skyler Carlson"]
packages = [
    { include = "app" },
    { include = "bot" }
]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.14"
orjson = "^3.10.16"
pydantic = "^2.11.3"
requests = "^2.32.3"
uvicorn = "^0.35.0"
pytest = "^8.3.5"
unittest2 = "^1.1.0"
loguru = "^0.7.3"
tenacity = "^9.1.2"
pre-commit = "^4.2.0"
httpx = "^0.28.1"
pytest-env = "^1.1.5"
pandas = "^2.2.3"
toml = "^0.10.2"
boto3 = "^1.37.32"
python-telegram-bot = "^21.11.1"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^1.0.0"
ruff = "^0.12.1"
